+++ OpenDentBusiness/Data Interface/LabTurnarounds.cs
@@ -44,7 +44,7 @@ namespace OpenDentBusiness{
 		}
 
 		///<summary></summary>
-		public static int Insert(LabTurnaround lab){
+		public static long Insert(LabTurnaround lab) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
 				lab.LabTurnaroundNum=Meth.GetInt(MethodBase.GetCurrentMethod(),lab);
 				return lab.LabTurnaroundNum;
