+++ OpenDentBusiness/TableTypes/PatPlan.cs
@@ -12,7 +12,7 @@ namespace OpenDentBusiness{
 		///<summary>FK to insplan.PlanNum.  The insurance plan attached to the patient.</summary>
 		public long PlanNum;
 		///<summary>Number like 1, 2, 3, etc.  Represents primary ins, secondary ins, tertiary ins, etc. 0 is not used</summary>
-		public long Ordinal;
+		public int Ordinal;
 		///<summary>For informational purposes only. For now, we lose the previous feature which let us set isPending without entering a plan.  Now, you have to enter the plan in order to check this box.</summary>
 		public bool IsPending;
 		///<summary>Enum:Relat Remember that this may need to be changed in the Claim also, if already created.</summary>
